pyproject.toml
[project]
name = "myth-weaver"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "rollforge",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "ollama>=0.1.7",
    "click>=8.3.1",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[tool.uv.sources]
rollforge = { path = "wheels/rollforge-0.1.0-cp312-cp312-win_amd64.whl" }
